requirements.txt
asgiref==3.5.2
Django==4.0.6
Pillow==9.3.0
sqlparse==0.4.2
tzdata==2022.1
